Extract review handler logging helpers and mention write-path utilities.
Move plan/doctrine snapshots, reducer and candidate-finding logs, diff hunk embedding, commit messages, mention write keys, and PR review diff collection into focused modules.
