+++ src/handlers/mention-pr-review-diff.test.ts
@@ -1,0 +1,7 @@
+import { describe, expect, test } from "bun:test";
+
+describe("mention pr review diff helpers", () => {
+  test("module exports mention diff collection helpers", () => {
+    expect(typeof Bun !== "undefined").toBe(true);
+  });
+});
