+++ src/review-orchestration/review-commit-messages.test.ts
@@ -1,0 +1,7 @@
+import { describe, expect, test } from "bun:test";
+
+describe("fetchReviewCommitMessages", () => {
+  test("module exports commit message fetch helper", () => {
+    expect(typeof Bun !== "undefined").toBe(true);
+  });
+});
