+++ src/review-orchestration/review-commit-messages.ts
@@ -1,0 +1,24 @@
+import type { GitHubApp } from "../auth/github-app.ts";
+
+export async function fetchReviewCommitMessages(
+  octokit: Awaited<ReturnType<GitHubApp["getInstallationOctokit"]>>,
+  owner: string,
+  repo: string,
+  prNumber: number,
+  commitCount: number,
+): Promise<Array<{ sha: string; message: string }>> {
+  if (commitCount === 0) return [];
+
+  const perPage = Math.min(commitCount, 100);
+  const { data } = await octokit.rest.pulls.listCommits({
+    owner,
+    repo,
+    pull_number: prNumber,
+    per_page: perPage,
+  });
+
+  return data.map((c) => ({
+    sha: c.sha.slice(0, 7),
+    message: c.commit.message.split("\n")[0] ?? "",
+  }));
+}
