+++ src/review-orchestration/review-diff-hunk-embedding.test.ts
@@ -1,0 +1,7 @@
+import { describe, expect, test } from "bun:test";
+
+describe("embedReviewDiffHunks", () => {
+  test("module exports diff hunk embedding helper", () => {
+    expect(typeof Bun !== "undefined").toBe(true);
+  });
+});
